requirements.txt
openai==0.27.8
wikitextparser==0.53.0
mwparserfromhell==0.6.4
openai-whisper==20230314
google-cloud-speech==2.21.0
pysubs2==1.6.1
google-cloud-translate==3.12.0
google-cloud-storage==2.10.0
llama-cpp-python==0.2.11
